fixed compile warning
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS@4282 212acab6-be3b-0410-9dea-997c60f758d6

+++ trunk/Source/DialogTemplate.cpp
@@ -617,7 +617,7 @@ BYTE* CDialogTemplate::Save(DWORD& dwSize) {
     }
   }
 
-  assert(seeker - pbDlg == dwSize);
+  assert((DWORD) seeker - (DWORD) pbDlg == dwSize);
 
   // DONE!
   return pbDlg;
